refactor: Afternote 도메인 모델 및 레포지토리 리팩토링
- 패키지 경로 최적화 및 모델 구조 재구성 (com.kuit 제거 및 input 패키지 분리)
- AfternoteRepository 메서드 명칭 간소화 (getAfternoteDetail -> getDetail 등)
- 모델 명칭 변경 (PagedAfternotes -> ListPage, GetAfternotesInput -> GetListPageInput 등)
- UseCase 내 변경된 모델 및 메서드 반영
