--- feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/model/GetAfternotesInput.kt
@@ -1,8 +1,0 @@
-package com.afternote.feature.afternote.domain.model
-
-// AfternoteRepository.getAfternotes() 인자
-data class GetAfternotesInput(
-    val category: String?,
-    val page: Int,
-    val size: Int,
-)
