+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/model/input/CreateGalleryInput.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.domain.model
+package com.afternote.feature.afternote.domain.model.input
 
 data class CreateGalleryInput(
     val title: String,
