+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/model/input/GetListPageInput.kt
@@ -1,0 +1,8 @@
+package com.afternote.feature.afternote.domain.model.input
+
+// AfternoteRepository.getListPage() 인자
+data class GetListPageInput(
+    val category: String?,
+    val page: Int,
+    val size: Int,
+)
