+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/repository/AfternoteRepository.kt
@@ -1,12 +1,12 @@
-package com.kuit.afternote.feature.afternote.domain.repository
+package com.afternote.feature.afternote.domain.repository
 
-import com.afternote.feature.afternote.domain.model.CreateGalleryInput
-import com.afternote.feature.afternote.domain.model.CreateSocialInput
 import com.afternote.feature.afternote.domain.model.Detail
-import com.afternote.feature.afternote.domain.model.GetAfternotesInput
-import com.afternote.feature.afternote.domain.model.UpdateAfternoteInput
-import com.afternote.feature.afternote.domain.model.playlist.CreatePlaylistInput
-import com.kuit.afternote.feature.afternote.domain.model.PagedAfternotes
+import com.afternote.feature.afternote.domain.model.ListPage
+import com.afternote.feature.afternote.domain.model.input.CreateGalleryInput
+import com.afternote.feature.afternote.domain.model.input.CreatePlaylistInput
+import com.afternote.feature.afternote.domain.model.input.CreateSocialInput
+import com.afternote.feature.afternote.domain.model.input.GetListPageInput
+import com.afternote.feature.afternote.domain.model.input.UpdateInput
 
 /**
  * Afternote 도메인 Repository 인터페이스.
@@ -19,20 +19,20 @@ import com.kuit.afternote.feature.afternote.domain.model.PagedAfternotes
  * - DELETE /afternotes/{afternoteId} (삭제)
  */
 interface AfternoteRepository {
-    suspend fun getAfternotes(input: GetAfternotesInput): Result<PagedAfternotes>
+    suspend fun getListPage(input: GetListPageInput): Result<ListPage>
 
     suspend fun createSocial(input: CreateSocialInput): Result<Long>
 
     suspend fun createGallery(input: CreateGalleryInput): Result<Long>
 
-    suspend fun getAfternoteDetail(afternoteId: Long): Result<Detail>
+    suspend fun getDetail(id: Long): Result<Detail>
 
     suspend fun createPlaylist(input: CreatePlaylistInput): Result<Long>
 
-    suspend fun updateAfternote(
-        afternoteId: Long,
-        input: UpdateAfternoteInput,
+    suspend fun update(
+        id: Long,
+        input: UpdateInput,
     ): Result<Long>
 
-    suspend fun deleteAfternote(afternoteId: Long): Result<Unit>
+    suspend fun delete(id: Long): Result<Unit>
 }
