+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/repository/MusicSearchRepository.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.afternote.domain.repository
 
-import com.kuit.afternote.feature.afternote.domain.model.SearchedSong
+import com.afternote.feature.afternote.domain.model.playlist.SearchedSong
 
 /**
  * Music search by keyword (artist or title).
