+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/usecase/CreateGalleryAfternoteUseCase.kt
@@ -1,7 +1,7 @@
 package com.afternote.feature.afternote.domain.usecase
 
-import com.afternote.feature.afternote.domain.model.CreateGalleryInput
-import com.kuit.afternote.feature.afternote.domain.repository.AfternoteRepository
+import com.afternote.feature.afternote.domain.model.input.CreateGalleryInput
+import com.afternote.feature.afternote.domain.repository.AfternoteRepository
 import javax.inject.Inject
 
 /**
