+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/usecase/DeleteAfternoteUseCase.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.afternote.domain.usecase
 
-import com.kuit.afternote.feature.afternote.domain.repository.AfternoteRepository
+import com.afternote.feature.afternote.domain.repository.AfternoteRepository
 import javax.inject.Inject
 
 /**
@@ -13,5 +13,5 @@ class DeleteAfternoteUseCase
     constructor(
         private val repository: AfternoteRepository,
     ) {
-        suspend operator fun invoke(afternoteId: Long): Result<Unit> = repository.deleteAfternote(afternoteId = afternoteId)
+        suspend operator fun invoke(afternoteId: Long): Result<Unit> = repository.delete(id = afternoteId)
     }
