+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/usecase/GetAfternoteDetailUseCase.kt
@@ -1,11 +1,11 @@
 package com.afternote.feature.afternote.domain.usecase
 
-import com.afternote.core.domain.model.ReceiverListItem
+import com.afternote.core.domain.model.setting.ReceiverListItem
 import com.afternote.core.domain.usecase.GetReceiversUseCase
 import com.afternote.core.domain.usecase.GetUserIdUseCase
 import com.afternote.feature.afternote.domain.model.Detail
-import com.kuit.afternote.feature.afternote.domain.model.DetailReceiver
-import com.kuit.afternote.feature.afternote.domain.repository.AfternoteRepository
+import com.afternote.feature.afternote.domain.model.DetailReceiver
+import com.afternote.feature.afternote.domain.repository.AfternoteRepository
 import javax.inject.Inject
 
 /**
@@ -22,7 +22,7 @@ class GetAfternoteDetailUseCase
         private val getUserIdUseCase: GetUserIdUseCase,
     ) {
         suspend operator fun invoke(afternoteId: Long): Result<Detail> {
-            val detailResult = repository.getAfternoteDetail(afternoteId = afternoteId)
+            val detailResult = repository.getDetail(id = afternoteId)
             val detail = detailResult.getOrElse { return detailResult }
             val userId = getUserIdUseCase() ?: return detailResult
             val receiversListResult = getReceiversUseCase(userId = userId)
